move copy to clipboard to effect
