+++ packages/app/src/app/overmind/effects/browser.ts
@@ -45,6 +45,17 @@ export default {
       close: () => popup?.close(),
     };
   },
+  copyToClipboard: (str: string) => {
+    const el = document.createElement('textarea');
+    el.value = str;
+    el.setAttribute('readonly', '');
+    el.style.position = 'absolute';
+    el.style.left = '-9999px';
+    document.body.appendChild(el);
+    el.select();
+    document.execCommand('copy');
+    document.body.removeChild(el);
+  },
   waitForMessage<T>(type): Promise<T> {
     return new Promise(resolve => {
       window.addEventListener('message', function onMessage(event) {
