+++ packages/app/src/app/pages/Sandbox/Editor/Header/Collaborators/ButtonActions.tsx
@@ -5,27 +5,15 @@ import { Stack, Button } from '@codesandbox/components';
 import CheckIcon from 'react-icons/lib/md/check';
 import { useOvermind } from 'app/overmind';
 
-const copyToClipboard = (str: string) => {
-  const el = document.createElement('textarea');
-  el.value = str;
-  el.setAttribute('readonly', '');
-  el.style.position = 'absolute';
-  el.style.left = '-9999px';
-  document.body.appendChild(el);
-  el.select();
-  document.execCommand('copy');
-  document.body.removeChild(el);
-};
-
 export const ButtonActions = () => {
   const [linkCopied, setLinkCopied] = React.useState(false);
   const [mounted, setMounted] = React.useState(false);
-  const { actions } = useOvermind();
+  const { actions, effects } = useOvermind();
   const timeout = React.useRef(null);
   const copyLink = () => {
     setLinkCopied(true);
 
-    copyToClipboard(document.location.href);
+    effects.browser.copyToClipboard(document.location.href);
 
     if (timeout.current) {
       clearTimeout(timeout.current);
